backend/requirements.txt
fastapi[all]
uvicorn
python-multipart
torch
openai-whisper
groq
python-dotenv
